pyproject.toml
[project]
name = "tingjian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "flask-login>=0.6.3",
    "flask-socketio>=5.5.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "langchain-core>=0.3.29",
    "openai>=1.59.5",
    "pillow>=11.1.0",
    "pyjwt>=2.10.1",
    "ultralytics>=8.3.145",
    "uvicorn[standard]>=0.34.0",
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.0"
Pillow = "^10.0.0"
jinja2 = "^3.1.2"
openai = "^1.3.0"
python-jwt = "^4.0.0"
